+++ src/ResourceManager/Sql/Commands.Sql/Database Activation/Cmdlet/SuspendAzureSqlDatabase.cs
@@ -21,7 +21,7 @@ namespace Microsoft.Azure.Commands.Sql.DatabaseActivation.Cmdlet
     /// <summary>
     /// Cmdlet to pause a Azure SQL Data Warehouse database.
     /// </summary>
-    [Cmdlet(VerbsLifecycle.Suspend, "AzureRmSqlDatabase", SupportsShouldProcess = true)]
+    [Cmdlet(VerbsLifecycle.Suspend, "AzureRmSqlDatabase", SupportsShouldProcess = true), OutputType(typeof(AzureSqlDatabaseModel))]
     public class SuspendAzureSqlDatabase : AzureSqlDatabaseActivationCmdletBase
     {
         /// <summary>