+++ src/ResourceManager/ApplicationInsights/Commands.ApplicationInsights/ChangeLog.md
@@ -18,7 +18,9 @@
         - Additional information about change #1
 -->
 ## Current Release
-* Updated help files to include full parameter types.
+
+## Version 0.1.5
+* Updated help files to include full parameter types and correct input/output types.
 
 ## Version 0.1.4
 * Set minimum dependency of module to PowerShell 5.0